+++ releaseNotes.md
@@ -1,3 +1,8 @@
+10.1.1 / 2017-11-05
+===================
+- New: started Tree Garden web console
+- Fix: Fixed create command line tool
+
 10.1.0 / 2017-11-04
 ===================
 - Fix: parsing top level program class fix